Revise returned value to match problem
The problem asks for the SUM of all of the 0-9 pandigital numbers
who are especially divisible, not the count of them.

Thus, returning "6" is incorrect.  What should instead be returned
is the sum of the 6 0-9 pandigital numbers who are especially divisible.

So, take their sum, stored in a long since 10+ digit number,
and return that sum.